Improve performance of "## no extract maketext" processing.
Instead of iteratively stripping out "## no extract maketext" lines,
which causes two regular expression matches on an entire file per
iteration, remove all in a single pass.

In a heavy use of this module across many files, this has been shown to
reduce the time to build a single locale database to roughly 21% of what
it used to be, while still producing identical results.

cPanel case for reference: 104873